комментарии + некоторые исправления

+++ account-processing/src/main/java/org/jedi_bachelor/service/AccountService.java
@@ -1,8 +1,13 @@
 package org.jedi_bachelor.service;
 
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.jedi_bachelor.model.entities.Account;
+import org.jedi_bachelor.model.entities.AccountBookRelationShip;
+import org.jedi_bachelor.model.entities.AccountBookSpeedReading;
 import org.jedi_bachelor.model.entities.FriendRelationship;
+import org.jedi_bachelor.repository.AccountBookRelationShipRepository;
+import org.jedi_bachelor.repository.AccountBookSpeedReadingRepository;
 import org.jedi_bachelor.repository.AccountRepository;
 import org.jedi_bachelor.repository.FriendRelationshipRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,11 +21,17 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class AccountService {
     @Autowired
+    private final AccountBookService accountBookService;
+    @Autowired
     private final RatingService ratingService;
     @Autowired
     private final AccountRepository accountRepository;
     @Autowired
     private final FriendRelationshipRepository friendRelationshipRepository;
+    @Autowired
+    private final AccountBookRelationShipRepository accountBookRelationShipRepository;
+    @Autowired
+    private final AccountBookSpeedReadingRepository accountBookSpeedReadingRepository;
 
     /*
     public void updateUserRating(KafkaDtoMessage message) {
@@ -31,7 +42,9 @@ public class AccountService {
      */
 
     /*
-    Метод для выдачи списка друзей аккаунта по аккаунту
+    * Метод для выдачи списка друзей аккаунта по аккаунту
+    * @param account (Account) - аккаунт, у которого надо найти список FriendRelationShip
+    * @return result (List<FriendRelationShip) - список отношений аккаунта
      */
     public List<FriendRelationship> getFriendsOfAccount(Account account) {
         List<FriendRelationship> result1 = this.friendRelationshipRepository.findFriendRelationshipByFriend1(account);
@@ -45,7 +58,9 @@ public class AccountService {
     }
 
     /*
-    Метод для добавления нового друга
+    * Метод для добавления нового друга
+    * @param ourId (Long) - наш собственный id
+    * @param id (Long) - id аккаунт, который надо добавить в друзья
      */
     public Boolean addNewFriend(Long ourId, Long id) {
         Optional<Account> account1 = this.accountRepository.getAccountById(ourId);
@@ -65,6 +80,11 @@ public class AccountService {
         return false;
     }
 
+    /*
+    * Метод для выдачи аккаунта по ID
+    * @param id - id аккаунта
+    * @return аккаунт с таким ID, если есть, и Optional.empty() если нет
+     */
     public Optional<Account> getAccountById(Long id) {
         if(!accountRepository.existsById(id))
             return Optional.empty();
@@ -72,10 +92,67 @@ public class AccountService {
         return accountRepository.getAccountById(id);
     }
 
+    /*
+    * Метод удаления аккаунта по ID
+    * @param id (Long) - id аккаунт, который надо удалить
+     */
+    @Transactional
     public void deleteAccountById(Long id) {
-        accountRepository.deleteById(id);
+        if(this.accountRepository.existsById(id)) {
+            accountRepository.deleteById(id);
+
+            // Удаление из друзей
+            List<FriendRelationship> friendRelationshipList1 = this.friendRelationshipRepository.findFriendRelationshipByFriend1(
+                    this.accountRepository.getAccountById(id).get()
+            );
+            List<FriendRelationship> friendRelationshipList2 = this.friendRelationshipRepository.findFriendRelationshipByFriend2(
+                    this.accountRepository.getAccountById(id).get()
+            );
+            List<FriendRelationship> allFriends = new ArrayList<>();
+            allFriends.addAll(friendRelationshipList1);
+            allFriends.addAll(friendRelationshipList2);
+
+            for(FriendRelationship friend : allFriends) {
+                friendRelationshipRepository.delete(friend);
+            }
+
+            // Удаление отношений аккаунт-чтение книг
+            List<AccountBookRelationShip> accountBookRelationShips1 = this.accountBookRelationShipRepository.findByAccountId(
+                    id
+            );
+            List<AccountBookRelationShip> accountBookRelationShips2 = this.accountBookRelationShipRepository.findByAccountId(
+                    id
+            );
+            List<AccountBookRelationShip> accountBookRelationShips = new ArrayList<>();
+            accountBookRelationShips.addAll(accountBookRelationShips1);
+            accountBookRelationShips.addAll(accountBookRelationShips2);
+
+            for(AccountBookRelationShip relation : accountBookRelationShips) {
+                accountBookRelationShipRepository.delete(relation);
+            }
+
+            // Удаление отношений аккаунт-скорость чтения
+            List<AccountBookSpeedReading> accountBookSpeeds1 = this.accountBookSpeedReadingRepository.findByAccountId(
+                    id
+            );
+            List<AccountBookSpeedReading> accountBookSpeeds2 = this.accountBookSpeedReadingRepository.findByAccountId(
+                    id
+            );
+            List<AccountBookSpeedReading> accountBookSpeeds = new ArrayList<>();
+            accountBookSpeeds.addAll(accountBookSpeeds1);
+            accountBookSpeeds.addAll(accountBookSpeeds2);
+
+            for(AccountBookSpeedReading s : accountBookSpeeds) {
+                accountBookSpeedReadingRepository.delete(s);
+            }
+        }
     }
 
+    /*
+    * Метод для изменения аккаунта
+    * @param account (Account) - аккаунт с изменёнными данными, который надо сохранить
+    * @return true, если аккаунт обновлён, false если нет.
+     */
     public Boolean updateAccount(Account account) {
         if(!accountRepository.existsById(account.getId())) {
             return false;
@@ -86,6 +163,11 @@ public class AccountService {
         return true;
     }
 
+    /*
+    * Метод для добавления нового аккаунта
+    * @param account (Account) - аккаунт, который надо добавить
+    * @return true, если аккаунт добавлен, false если нет.
+     */
     public Boolean addNewAccount(Account account) {
         if(accountRepository.existsById(account.getId()) ||
                 accountRepository.existsByUsername(account.getUsername()) ||
@@ -96,6 +178,11 @@ public class AccountService {
         return true;
     }
 
+    /*
+    * Метод для выдачи аккаунта по логину
+    * @param login (String) - логин пользователя
+    * @return аккаунт с таким логиком
+     */
     public Optional<Account> getAccountByLogin(String login) {
         return accountRepository.getAccountByUsername(login);
     }
